requirements.txt
a2a-sdk[http-server]>=0.3.25
langgraph>=1.1.0
langchain-core>=1.2.0
langchain-openai>=1.1.0
uvicorn>=0.41.0
httpx>=0.28.0
fastapi>=0.115.0
